Create README.md
# application-of-tsp-drilling-a-pcb
It was the a mini project and the topic was to demonstrate an application of travelling salesman problem.
There are 3 functions, main() takes input for the dimensions of the rectangular pcb and generates a matrix having 0's and 1's .
1 indicates that a hole is to be drilled. 
getinfo() generates distance of one hole to every other hole and generates the cost adjacency matrix. 
tsp generates the shortest path from the starting hole back to it passing through all other holes.